refactor!: delete useGraphNodeManager
All four responsibilities are resolved rather than relocated:

- `getNode`/`nodeRefs`: `flags.pinned` reads nodeDataStore; live-node
  lookups use `canvasStore.currentGraph.getNodeById`.
- slot-array graft: reactivity is now class-side from construction.
- `node:slot-label:changed` reprojection: slot objects are reactive, so
  `slot.label = ...` is already a tracked write.
- layoutStore seeding: moves to the `LGraph.add`/`LGraph.remove`
  chokepoints, matching `LGraph.createReroute`.

The `configuringGraph` deferral is unnecessary: `add` seeds the entry with
whatever geometry the node has, and the `pos`/`size` setters already write
through, so `configure` updates it as real values land.

The `onNodeAdded` replay loop is gone with it. It re-fired node-added for
every pre-existing node on each graph switch, leaking notifications to
unrelated subscribers; per-graph bootstrap is solely
`layoutStore.initializeFromLiteGraph`. The ancestor guard in
useErrorClearingHooks stays — it is reachable by any interior add — but its
comment no longer cites the deleted replay.

Tests move to the code they exercise: error flags to useNodeErrorFlagSync,
widget-store reactivity to widgetValueStore.graphReactivity, layout seeding
to LGraph, slot reactivity to NodeSlot.
