+++ browser_tests/tests/propertiesPanel/errorsTabModeAware.spec.ts
@@ -801,10 +801,10 @@ test.describe('Errors tab - Mode-aware errors', { tag: '@ui' }, () => {
     test('Entering a bypassed subgraph does not resurface interior missing model error', async ({
       comfyPage
     }) => {
-      // Regression: useGraphNodeManager replays graph.onNodeAdded for
-      // each interior node on subgraph entry; without an ancestor-aware
-      // guard in scanSingleNodeErrors, that re-scan reintroduced the
-      // error that the initial pipeline had correctly suppressed.
+      // Regression: entering a bypassed subgraph re-scans its interior
+      // nodes; without an ancestor-aware guard in scanSingleNodeErrors,
+      // that re-scan reintroduced the error the initial pipeline had
+      // correctly suppressed.
       await comfyPage.workflow.loadWorkflow(
         'missing/missing_models_in_bypassed_subgraph'
       )
