+++ src/composables/graph/useVueNodeLifecycle.ts
@@ -1,8 +1,6 @@
 import { createSharedComposable, whenever } from '@vueuse/core'
 import { shallowRef, watch } from 'vue'
 
-import { useGraphNodeManager } from '@/composables/graph/useGraphNodeManager'
-import type { GraphNodeManager } from '@/composables/graph/useGraphNodeManager'
 import { useVueFeatureFlags } from '@/composables/useVueFeatureFlags'
 import type { LGraphNode } from '@/lib/litegraph/src/litegraph'
 import { useCanvasStore } from '@/renderer/core/canvas/canvasStore'
@@ -11,23 +9,25 @@ import { layoutStore } from '@/renderer/core/layout/store/layoutStore'
 import { useLayoutSync } from '@/renderer/core/layout/sync/useLayoutSync'
 import { app as comfyApp } from '@/scripts/app'
 
+/**
+ * Bootstraps the renderer's layout view of the active graph and owns the
+ * Layout↔LiteGraph sync lifecycle. Per-node layout entries are seeded by
+ * `LGraph.add` / `LGraph.remove`; this re-seeds the whole active graph on
+ * entry, which is what makes subgraph navigation show the right nodes.
+ */
 function useVueNodeLifecycleIndividual() {
   const canvasStore = useCanvasStore()
   const layoutMutations = useLayoutMutations()
   const { shouldRenderVueNodes } = useVueFeatureFlags()
-  const nodeManager = shallowRef<GraphNodeManager | null>(null)
+  const isInitialized = shallowRef(false)
   const { startSync, stopSync } = useLayoutSync()
 
-  const initializeNodeManager = () => {
+  const initializeVueNodeLayout = () => {
     // Use canvas graph if available (handles subgraph contexts), fallback to app graph
     const activeGraph = comfyApp.canvas?.graph
-    if (!activeGraph || nodeManager.value) return
+    if (!activeGraph || isInitialized.value) return
+    isInitialized.value = true
 
-    // Initialize the core node manager
-    const manager = useGraphNodeManager(activeGraph)
-    nodeManager.value = manager
-
-    // Initialize layout system with existing nodes from active graph
     const nodes = activeGraph._nodes.map((node: LGraphNode) => ({
       id: node.id,
       pos: [node.pos[0], node.pos[1]] as [number, number],
@@ -35,7 +35,8 @@ function useVueNodeLifecycleIndividual() {
     }))
     layoutStore.initializeFromLiteGraph(nodes)
 
-    // Seed reroutes into the Layout Store so hit-testing uses the new path
+    // Deserialised reroutes bypass createReroute, so they need seeding here for
+    // hit-testing to find them.
     for (const reroute of activeGraph.reroutes.values()) {
       const [x, y] = reroute.pos
       layoutMutations.createReroute(reroute.id, { x, y })
@@ -46,16 +47,9 @@ function useVueNodeLifecycleIndividual() {
     startSync(canvasStore.canvas)
   }
 
-  const disposeNodeManagerAndSyncs = () => {
+  const disposeVueNodeLayout = () => {
     stopSync()
-    if (!nodeManager.value) return
-
-    try {
-      nodeManager.value.cleanup()
-    } catch {
-      /* empty */
-    }
-    nodeManager.value = null
+    isInitialized.value = false
   }
 
   // Watch for Vue nodes enabled state changes
@@ -63,7 +57,7 @@ function useVueNodeLifecycleIndividual() {
     () => shouldRenderVueNodes.value && Boolean(comfyApp.canvas?.graph),
     (enabled) => {
       if (enabled) {
-        initializeNodeManager()
+        initializeVueNodeLayout()
       }
     },
     { immediate: true }
@@ -72,7 +66,7 @@ function useVueNodeLifecycleIndividual() {
   whenever(
     () => !shouldRenderVueNodes.value,
     () => {
-      disposeNodeManagerAndSyncs()
+      disposeVueNodeLayout()
 
       // Force arrange() on all nodes so input.pos is computed before
       // the first legacy drawConnections frame (which may run before
@@ -103,49 +97,34 @@ function useVueNodeLifecycleIndividual() {
     }
   )
 
-  // Handle case where Vue nodes are enabled but graph starts empty
+  /**
+   * When the canvas mounts before its first node exists, bootstrap is deferred
+   * to the first add. The last onNodeAdded monkeypatch in the renderer; it goes
+   * away once LGraph dispatches a `node:added` event.
+   */
   const setupEmptyGraphListener = () => {
     const activeGraph = comfyApp.canvas?.graph
     if (
       !shouldRenderVueNodes.value ||
-      nodeManager.value ||
+      isInitialized.value ||
       activeGraph?._nodes.length !== 0
     ) {
       return
     }
     const originalOnNodeAdded = activeGraph.onNodeAdded
     activeGraph.onNodeAdded = function (node: LGraphNode) {
-      // Restore original handler
       activeGraph.onNodeAdded = originalOnNodeAdded
 
-      // Initialize node manager if needed
-      if (shouldRenderVueNodes.value && !nodeManager.value) {
-        initializeNodeManager()
-      }
+      if (shouldRenderVueNodes.value) initializeVueNodeLayout()
 
-      // Call original handler
-      if (originalOnNodeAdded) {
-        originalOnNodeAdded.call(this, node)
-      }
-    }
-  }
-
-  // Cleanup function for component unmounting
-  const cleanup = () => {
-    if (nodeManager.value) {
-      nodeManager.value.cleanup()
-      nodeManager.value = null
+      originalOnNodeAdded?.call(this, node)
     }
   }
 
   return {
-    nodeManager,
-
-    // Lifecycle methods
-    initializeNodeManager,
-    disposeNodeManagerAndSyncs,
-    setupEmptyGraphListener,
-    cleanup
+    initializeVueNodeLayout,
+    disposeVueNodeLayout,
+    setupEmptyGraphListener
   }
 }
