+++ src/lib/litegraph/src/LGraphCanvas.ghost.test.ts
@@ -8,6 +8,9 @@ import { createMockCanvasRenderingContext2D } from '@/utils/__tests__/litegraphT
 
 vi.mock('@/renderer/core/layout/store/layoutStore', () => ({
   layoutStore: {
+    getCurrentSource: vi.fn(),
+    getCurrentActor: vi.fn(),
+    applyOperation: vi.fn(),
     querySlotAtPoint: vi.fn(),
     queryRerouteAtPoint: vi.fn(),
     getNodeLayoutRef: vi.fn(() => ({ value: null })),
