+++ src/lib/litegraph/src/LGraphCanvas.slotHitDetection.test.ts
@@ -12,6 +12,9 @@ import { layoutStore } from '@/renderer/core/layout/store/layoutStore'
 
 vi.mock('@/renderer/core/layout/store/layoutStore', () => ({
   layoutStore: {
+    getCurrentSource: vi.fn(),
+    getCurrentActor: vi.fn(),
+    applyOperation: vi.fn(),
     querySlotAtPoint: vi.fn(),
     queryRerouteAtPoint: vi.fn(),
     getNodeLayoutRef: vi.fn(() => ({ value: null })),
