+++ src/renderer/extensions/vueNodes/composables/useNodeEventHandlers.test.ts
@@ -1,16 +1,22 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { computed, shallowRef } from 'vue'
+import { computed } from 'vue'
 
-import { useGraphNodeManager } from '@/composables/graph/useGraphNodeManager'
-import type { GraphNodeManager } from '@/composables/graph/useGraphNodeManager'
-import { useVueNodeLifecycle } from '@/composables/graph/useVueNodeLifecycle'
-import type { LGraphCanvas, LGraphNode } from '@/lib/litegraph/src/litegraph'
+import type {
+  LGraph,
+  LGraphCanvas,
+  LGraphNode
+} from '@/lib/litegraph/src/litegraph'
 import { useCanvasStore } from '@/renderer/core/canvas/canvasStore'
 import { useLayoutMutations } from '@/renderer/core/layout/operations/layoutMutations'
 import { useNodeEventHandlers } from '@/renderer/extensions/vueNodes/composables/useNodeEventHandlers'
 import { toNodeId } from '@/types/nodeId'
 
 const canvasSelectedItems = vi.hoisted(() => [] as Array<{ id?: string }>)
+const graphNode = vi.hoisted(() => ({
+  id: 'node-1',
+  selected: false,
+  flags: { pinned: false }
+}))
 
 vi.mock('@/renderer/core/canvas/canvasStore', () => {
   const canvas: Partial<LGraphCanvas> = {
@@ -19,8 +25,12 @@ vi.mock('@/renderer/core/canvas/canvasStore', () => {
     deselectAll: vi.fn()
   }
   const updateSelectedItems = vi.fn()
+  const currentGraph: Partial<LGraph> = {
+    getNodeById: vi.fn(() => graphNode as Partial<LGraphNode> as LGraphNode)
+  }
   const canvasStoreInstance = {
     canvas: canvas as LGraphCanvas,
+    currentGraph: currentGraph as LGraph,
     updateSelectedItems,
     selectedItems: canvasSelectedItems
   }
@@ -46,33 +56,8 @@ vi.mock('@/renderer/core/layout/operations/layoutMutations', () => {
   }
 })
 
-vi.mock('@/composables/graph/useGraphNodeManager', () => {
-  const mockNode = {
-    id: 'node-1',
-    selected: false,
-    flags: { pinned: false }
-  }
-  const nodeManager = shallowRef({
-    getNode: vi.fn(() => mockNode as Partial<LGraphNode> as LGraphNode)
-  } as Partial<GraphNodeManager> as GraphNodeManager)
-  return {
-    useGraphNodeManager: vi.fn(() => nodeManager)
-  }
-})
-
-vi.mock('@/composables/graph/useVueNodeLifecycle', () => {
-  const nodeManager = useGraphNodeManager(null!)
-  return {
-    useVueNodeLifecycle: vi.fn(() => ({
-      nodeManager
-    }))
-  }
-})
-
 describe('useNodeEventHandlers', () => {
-  const { nodeManager: mockNodeManager } = useVueNodeLifecycle()
-
-  const mockNode = mockNodeManager.value!.getNode(toNodeId('fake_id'))
+  const mockNode = graphNode as Partial<LGraphNode> as LGraphNode
   const mockLayoutMutations = useLayoutMutations()
 
   const testNodeId = toNodeId('node-1')
